Refactoring of spring
