+++ particlefgen.cpp
@@ -21,3 +21,53 @@ void particleDrag::updateForce(particle *particle, float dT) {
 	force *= -dragCoeff;
 	particle->addForce(force);
 }
+
+void SpringParticle::updateForce(particle* particle, float duration) {
+	glm::vec3 force;
+
+	force = particle->getPosition();
+	force -= other->getPosition();
+
+	float magnitude = force.length();
+
+	magnitude = abs(magnitude - restLength);
+	magnitude *= springConstant;
+
+	force = glm::normalize(force);
+	force *= magnitude;
+	particle->addForce(force);
+}
+
+void AnchoredSpring::updateForce(particle* particle, float duration) {
+	glm::vec3 force;
+
+	force = particle->getPosition();
+	force -= *anchor;
+
+	float magnitude = force.length();
+
+	magnitude = abs(magnitude - restLength);
+	magnitude *= springConstant;
+
+	force = glm::normalize(force);
+	force *= magnitude;
+	particle->addForce(force);
+}
+
+void ElasticBungee::updateForce(particle* particle, float duration) {
+	glm::vec3 force;
+
+	force = particle->getPosition();
+	force -= other->getPosition();
+
+	float magnitude = force.length();
+
+	magnitude = abs(magnitude - restLength);
+	if (magnitude <= restLength) return;
+
+	magnitude = springConstant * (restLength - magnitude);
+
+	force = glm::normalize(force);
+	force *= magnitude;
+	particle->addForce(force);
+}
